refactor(errors): split errors.ts into multiple files
Split the large errors.ts file (996 lines) into multiple smaller files organized by
error category to improve maintainability.

- Created modular directory structure for errors
- Separated base error class and error categories
- Grouped specialized error classes by domain
- Organized factory functions by domain
- Maintained backward compatibility through index.ts
- Ensured all tests pass with the new structure

Resolves AC 2.2 (Large file size of errors.ts)
